@@ -1,6 +1,6 @@
 {
   "name": "system-management-demo",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -13,47 +13,48 @@
     "french": "ng build --prod --i18n-file src/locale/messages.fr.xlf --i18n-format xlf --i18n-locale fr",
     "spanish": "ng build --prod"
   },
+  "description": "System Management Demo is a proyect to management new users and countries,  applying all standars.",
+  "author": "Suarez Molina Jesus",
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.0.0",
-    "@angular/cdk": "~7.1.1",
-    "@angular/common": "~7.0.0",
-    "@angular/compiler": "~7.0.0",
-    "@angular/core": "~7.0.0",
+    "@angular/animations": "~8.0.3",
+    "@angular/cdk": "~8.0.1",
+    "@angular/common": "~8.0.3",
+    "@angular/compiler": "~8.0.3",
+    "@angular/core": "~8.0.3",
     "@angular/fire": "^5.1.1",
-    "@angular/flex-layout": "^7.0.0-beta.19",
-    "@angular/forms": "~7.0.0",
-    "@angular/http": "~7.0.0",
-    "@angular/material": "^7.1.1",
-    "@angular/platform-browser": "~7.0.0",
-    "@angular/platform-browser-dynamic": "~7.0.0",
-    "@angular/router": "~7.0.0",
-    "angularfire2": "^5.1.1",
-    "core-js": "^2.5.4",
-    "firebase": "^5.7.0",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "~8.0.3",
+    "@angular/material": "^8.0.1",
+    "@angular/platform-browser": "~8.0.3",
+    "@angular/platform-browser-dynamic": "~8.0.3",
+    "@angular/router": "~8.0.3",
+    "core-js": "^3.1.4",
+    "firebase": "5.10.1",
     "hammerjs": "^2.0.8",
-    "rxjs": "~6.3.3",
-    "zone.js": "~0.8.26"
+    "rxjs": "~6.5.2",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "~7.0.6",
-    "@angular/compiler-cli": "~7.0.0",
-    "@angular/language-service": "~7.0.0",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/cli": "~8.0.4",
+    "@angular/compiler-cli": "~8.0.3",
+    "@angular/language-service": "~8.0.3",
+    "@types/node": "~12.0.10",
+    "@types/jasmine": "~3.3.13",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "ts-node": "~8.3.0",
+    "tslint": "~5.18.0",
+    "typescript": "~3.4.3"
   }
 }
