@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
-import { HttpModule } from '@angular/http';
 import { PaisesApiService } from './paises-api.service';
 import { ApiModule } from '../api.module';
 
@@ -9,7 +8,7 @@ import { ApiModule } from '../api.module';
   declarations: [],
   imports: [
     HttpClientModule,
-    HttpModule,
+    HttpClientModule,
     CommonModule,
     ApiModule
   ],
